+++ packages/reference_example/android/src/main/java/com/example/reference_example/classnameProxy.java
@@ -1,0 +1,23 @@
+package com.example.reference_example;
+
+import com.example.reference_example.fakelibrary.ClassTemplate;
+
+public class classnameProxy implements LibraryTemplate.$__class_name__ {
+  public final ClassTemplate classTemplate;
+
+  public static Double staticMethodTemplate(String parameterTemplate) {
+    return ClassTemplate.staticMethodTemplate(parameterTemplate);
+  }
+
+  public classnameProxy(Integer fieldTemplate) {
+    this(new ClassTemplate(fieldTemplate));
+  }
+
+  public classnameProxy(ClassTemplate classTemplate) {
+    this.classTemplate = classTemplate;
+  }
+
+  public String methodTemplate(String parameterTemplate) {
+    return classTemplate.methodTemplate(parameterTemplate);
+  }
+}