+++ packages/reference_example/ios/Classes/EXPChannelRegistrar.m
@@ -1,13 +1,13 @@
 #import "EXPChannelRegistrar.h"
 
 @implementation EXPLibraryImplementations
-- (nonnull EXPClassTemplateHandler *)classTemplateHandler {
+- (nonnull EXPClassTemplateHandler *)handler__class_name__ {
   return [[EXPClassTemplateHandler alloc] init];
 }
 @end
 
 @implementation EXPClassTemplateProxy
-+(NSNumber *)staticMethodTemplate:(NSString *)parameterTemplate {
++ (NSNumber *)staticMethodTemplate:(NSString *)parameterTemplate {
   return @([ClassTemplate staticMethodTemplate:parameterTemplate]);
 }
 
@@ -24,21 +24,25 @@
   return self;
 }
 
-- (NSNumber * _Nullable)fieldTemplate {
-  return nil;
-}
-
 - (NSString *_Nullable)methodTemplate:(NSString *_Nullable)parameterTemplate {
   return [_classTemplate methodTemplate:parameterTemplate];
 }
 @end
 
 @implementation EXPClassTemplateHandler
-- (NSObject<REFClassTemplate> *)onCreate:(REFTypeChannelMessenger *)messenger args:(REFClassTemplateCreationArgs *)args {
-  return [[EXPClassTemplateProxy alloc] initWithFieldTemplate:args.fieldTemplate];
+- (NSObject<__prefix____class_name__> *)__create:(REFTypeChannelMessenger *)messenger
+                                  __field_name__:(NSNumber *)__field_name__ {
+  return [[EXPClassTemplateProxy alloc] initWithFieldTemplate:__field_name__];
 }
 
-- (NSNumber *)on_staticMethodTemplate:(REFTypeChannelMessenger *)messenger parameterTemplate:(NSString *)parameterTemplate {
-  return [EXPClassTemplateProxy staticMethodTemplate:parameterTemplate];
+- (NSObject *)___staticMethod_name__:(REFTypeChannelMessenger *)messenger
+                  __parameter_name__:(NSString *)__parameter_name__ {
+  return [EXPClassTemplateProxy staticMethodTemplate:__parameter_name__];
+}
+
+- (NSObject *)___method_name__:(NSObject<__prefix____class_name__> *)_instance
+            __parameter_name__:(NSString *)__parameter_name__ {
+  EXPClassTemplateProxy *proxy = (EXPClassTemplateProxy *)_instance;
+  return [proxy methodTemplate:__parameter_name__];
 }
 @end